Creater base_chart.rb file in googlecharts
